+++ mithril-test-lab/mithril-end-to-end/src/toolkit/check/mod.rs
@@ -1,3 +1,19 @@
+#![allow(unused_imports)]
+
+mod cardano_blocks_transactions;
+mod cardano_database;
+mod cardano_stake_distribution;
+mod cardano_transactions;
+mod certificate;
+mod mithril_stake_distribution;
+
+pub use cardano_blocks_transactions::*;
+pub use cardano_database::*;
+pub use cardano_stake_distribution::*;
+pub use cardano_transactions::*;
+pub use certificate::*;
+pub use mithril_stake_distribution::*;
+
 use std::{path::PathBuf, time::Duration};
 
 use anyhow::{Context, anyhow};
@@ -41,806 +57,26 @@ async fn get_json_response<T: DeserializeOwned>(url: String) -> StdResult<reqwes
 
 #[derive(Debug, Clone, Default)]
 pub struct CheckToolkit {
-    context: ScenarioToolkitContext,
+    pub cardano_blocks_transactions: CheckCardanoBlocksTransactionsToolkit,
+    pub cardano_database: CheckCardanoDatabaseToolkit,
+    pub cardano_stake_distribution: CheckCardanoStakeDistributionToolkit,
+    pub cardano_transactions: CheckCardanoTransactionsToolkit,
+    pub certificate: CheckCertificateToolkit,
+    pub mithril_stake_distribution: CheckMithrilStakeDistributionToolkit,
 }
 
 impl CheckToolkit {
     pub fn new(context: ScenarioToolkitContext) -> Self {
-        Self { context }
-    }
-
-    pub async fn node_producing_mithril_stake_distribution(
-        &self,
-        aggregator: &Aggregator,
-    ) -> StdResult<String> {
-        let url = format!(
-            "{}/artifact/mithril-stake-distributions",
-            aggregator.endpoint()
-        );
-        info!("Waiting for the aggregator to produce a mithril stake distribution"; "aggregator" => &aggregator.name());
-
-        async fn fetch_last_mithril_stake_distribution_hash(
-            url: String,
-        ) -> StdResult<Option<String>> {
-            match get_json_response::<MithrilStakeDistributionListMessage>(url)
-                .await?
-                .as_deref()
-            {
-                Ok([stake_distribution, ..]) => Ok(Some(stake_distribution.hash.clone())),
-                Ok(&[]) => Ok(None),
-                Err(err) => Err(anyhow!("Invalid mithril stake distribution body: {err}",)),
-            }
+        Self {
+            cardano_blocks_transactions: CheckCardanoBlocksTransactionsToolkit::new(
+                context.clone(),
+            ),
+            cardano_database: CheckCardanoDatabaseToolkit::new(context.clone()),
+            cardano_stake_distribution: CheckCardanoStakeDistributionToolkit::new(context.clone()),
+            cardano_transactions: CheckCardanoTransactionsToolkit::new(context.clone()),
+            certificate: CheckCertificateToolkit::new(context.clone()),
+            mithril_stake_distribution: CheckMithrilStakeDistributionToolkit::new(context.clone()),
         }
-
-        match attempt!(30, Duration::from_secs(3), {
-        fetch_last_mithril_stake_distribution_hash(url.clone()).await
-    }) {
-            AttemptResult::Ok(hash) => {
-                info!("Aggregator produced a mithril stake distribution"; "hash" => &hash, "aggregator" => &aggregator.name());
-                Ok(hash)
-            }
-            AttemptResult::Err(error) => Err(error),
-            AttemptResult::Timeout() => Err(anyhow!(
-            "Timeout exhausted assert_node_producing_mithril_stake_distribution, no response from `{url}`"
-        )),
-        }.with_context(|| {
-            format!(
-                "Requesting aggregator `{}`",
-                aggregator.name()
-            )
-        })
-    }
-
-    pub async fn signer_is_signing_mithril_stake_distribution(
-        &self,
-        aggregator: &Aggregator,
-        hash: &str,
-        expected_epoch_min: Epoch,
-    ) -> StdResult<String> {
-        let url = format!(
-            "{}/artifact/mithril-stake-distribution/{hash}",
-            aggregator.endpoint()
-        );
-        info!(
-            "Asserting the aggregator is signing the mithril stake distribution message `{}` with an expected min epoch of `{}`",
-            hash,
-            expected_epoch_min;
-            "aggregator" => &aggregator.name()
-        );
-
-        async fn fetch_mithril_stake_distribution_message(
-            url: String,
-            expected_epoch_min: Epoch,
-        ) -> StdResult<Option<MithrilStakeDistributionMessage>> {
-            match get_json_response::<MithrilStakeDistributionMessage>(url.clone()).await? {
-                Ok(stake_distribution) => match stake_distribution.epoch {
-                    epoch if epoch >= expected_epoch_min => Ok(Some(stake_distribution)),
-                    epoch => Err(anyhow!(
-                        "Minimum expected mithril stake distribution epoch not reached: {epoch} < {expected_epoch_min}"
-                    )),
-                },
-                Err(err) => Err(anyhow!("Invalid mithril stake distribution body: {err}",)),
-            }
-        }
-
-        match attempt!(10, Duration::from_millis(1000), {
-        fetch_mithril_stake_distribution_message(url.clone(), expected_epoch_min).await
-    }) {
-            AttemptResult::Ok(stake_distribution) => {
-                info!("Signer signed a mithril stake distribution"; "certificate_hash" => &stake_distribution.certificate_hash, "aggregator" => &aggregator.name());
-                Ok(stake_distribution.certificate_hash)
-            }
-            AttemptResult::Err(error) => Err(error),
-            AttemptResult::Timeout() => Err(anyhow!(
-            "Timeout exhausted assert_signer_is_signing_mithril_stake_distribution, no response from `{url}`"
-        )),
-        }.with_context(|| {
-            format!(
-                "Requesting aggregator `{}`",
-                aggregator.name()
-            )
-        })
-    }
-
-    pub async fn node_producing_cardano_database_snapshot(
-        &self,
-        aggregator: &Aggregator,
-    ) -> StdResult<String> {
-        let url = format!("{}/artifact/cardano-database", aggregator.endpoint());
-        info!("Waiting for the aggregator to produce a Cardano database snapshot"; "aggregator" => &aggregator.name());
-
-        async fn fetch_last_cardano_database_snapshot_hash(
-            url: String,
-        ) -> StdResult<Option<String>> {
-            match get_json_response::<CardanoDatabaseSnapshotListMessage>(url)
-                .await?
-                .as_deref()
-            {
-                Ok([cardano_database_snapshot, ..]) => {
-                    Ok(Some(cardano_database_snapshot.hash.clone()))
-                }
-                Ok(&[]) => Ok(None),
-                Err(err) => Err(anyhow!("Invalid Cardano database snapshot body: {err}",)),
-            }
-        }
-
-        match attempt!(30, Duration::from_millis(2000), {
-        fetch_last_cardano_database_snapshot_hash(url.clone()).await
-    }) {
-            AttemptResult::Ok(hash) => {
-                info!("Aggregator produced a Cardano database snapshot"; "hash" => &hash, "aggregator" => &aggregator.name());
-                Ok(hash)
-            }
-            AttemptResult::Err(error) => Err(error),
-            AttemptResult::Timeout() => Err(anyhow!(
-            "Timeout exhausted assert_node_producing_snapshot, no response from `{url}`"
-        )),
-        }
-            .with_context(|| format!("Requesting aggregator `{}`", aggregator.name()))
-    }
-
-    pub async fn signer_is_signing_cardano_database_snapshot(
-        &self,
-        aggregator: &Aggregator,
-        hash: &str,
-        expected_epoch_min: Epoch,
-    ) -> StdResult<String> {
-        let url = format!("{}/artifact/cardano-database/{hash}", aggregator.endpoint());
-        info!(
-            "Asserting the aggregator is signing the Cardano database snapshot message `{}` with an expected min epoch of `{}`",
-            hash,
-            expected_epoch_min;
-            "aggregator" => &aggregator.name()
-        );
-
-        async fn fetch_cardano_database_snapshot_message(
-            url: String,
-            expected_epoch_min: Epoch,
-        ) -> StdResult<Option<CardanoDatabaseSnapshotMessage>> {
-            match get_json_response::<CardanoDatabaseSnapshotMessage>(url).await? {
-                Ok(cardano_database_snapshot) => match cardano_database_snapshot.beacon.epoch {
-                    epoch if epoch >= expected_epoch_min => Ok(Some(cardano_database_snapshot)),
-                    epoch => Err(anyhow!(
-                        "Minimum expected Cardano database snapshot epoch not reached: {epoch} < {expected_epoch_min}"
-                    )),
-                },
-                Err(err) => Err(anyhow!(err).context("Invalid Cardano database snapshot body")),
-            }
-        }
-
-        match attempt!(10, Duration::from_millis(1000), {
-        fetch_cardano_database_snapshot_message(url.clone(), expected_epoch_min).await
-    }) {
-            AttemptResult::Ok(snapshot) => {
-                info!("Signer signed a snapshot"; "certificate_hash" => &snapshot.certificate_hash, "aggregator" => &aggregator.name());
-                Ok(snapshot.certificate_hash)
-            }
-            AttemptResult::Err(error) => Err(error),
-            AttemptResult::Timeout() => Err(anyhow!(
-            "Timeout exhausted assert_signer_is_signing_snapshot, no response from `{url}`"
-        )),
-        }
-            .with_context(|| format!("Requesting aggregator `{}`", aggregator.name()))
-    }
-
-    pub async fn node_producing_cardano_database_digests_map(
-        &self,
-        aggregator: &Aggregator,
-    ) -> StdResult<Vec<(String, String)>> {
-        let url = format!(
-            "{}/artifact/cardano-database/digests",
-            aggregator.endpoint()
-        );
-        info!("Waiting for the aggregator to produce a Cardano database digests map"; "aggregator" => &aggregator.name());
-
-        async fn fetch_cardano_database_digests_map(
-            url: String,
-        ) -> StdResult<Option<Vec<(String, String)>>> {
-            match get_json_response::<CardanoDatabaseDigestListMessage>(url)
-                .await?
-                .as_deref()
-            {
-                Ok(&[]) => Ok(None),
-                Ok(cardano_database_digests_map) => Ok(Some(
-                    cardano_database_digests_map
-                        .iter()
-                        .map(|item| (item.immutable_file_name.clone(), item.digest.clone()))
-                        .collect(),
-                )),
-                Err(err) => Err(anyhow!("Invalid Cardano database digests map body: {err}",)),
-            }
-        }
-
-        match attempt!(30, Duration::from_millis(2000), {
-        fetch_cardano_database_digests_map(url.clone()).await
-    }) {
-            AttemptResult::Ok(cardano_database_digests_map) => {
-                info!("Aggregator produced a Cardano database digests map"; "total_digests" => &cardano_database_digests_map.len(), "aggregator" => &aggregator.name());
-                Ok(cardano_database_digests_map)
-            }
-            AttemptResult::Err(error) => Err(error),
-            AttemptResult::Timeout() => Err(anyhow!(
-            "Timeout exhausted assert_node_producing_cardano_database_digests_map, no response from `{url}`"
-        )),
-        }.with_context(|| {
-            format!(
-                "Requesting aggregator `{}`",
-                aggregator.name()
-            )
-        })
-    }
-
-    pub async fn node_producing_cardano_transactions(
-        &self,
-        aggregator: &Aggregator,
-    ) -> StdResult<String> {
-        let url = format!("{}/artifact/cardano-transactions", aggregator.endpoint());
-        info!("Waiting for the aggregator to produce a Cardano transactions artifact"; "aggregator" => &aggregator.name(), "aggregator" => &aggregator.name());
-
-        async fn fetch_last_cardano_transaction_snapshot_hash(
-            url: String,
-        ) -> StdResult<Option<String>> {
-            match get_json_response::<CardanoTransactionSnapshotListMessage>(url)
-                .await?
-                .as_deref()
-            {
-                Ok([artifact, ..]) => Ok(Some(artifact.hash.clone())),
-                Ok(&[]) => Ok(None),
-                Err(err) => Err(anyhow!("Invalid Cardano transactions artifact body: {err}",)),
-            }
-        }
-
-        match attempt!(30, Duration::from_millis(2000), {
-        fetch_last_cardano_transaction_snapshot_hash(url.clone()).await
-    }) {
-            AttemptResult::Ok(hash) => {
-                info!("Aggregator produced a Cardano transactions artifact"; "hash" => &hash, "aggregator" => &aggregator.name());
-                Ok(hash)
-            }
-            AttemptResult::Err(error) => Err(error),
-            AttemptResult::Timeout() => Err(anyhow!(
-            "Timeout exhausted assert_node_producing_cardano_transactions, no response from `{url}`"
-        )),
-        }
-            .with_context(|| format!("Requesting aggregator `{}`", aggregator.name()))
-    }
-
-    pub async fn signer_is_signing_cardano_transactions(
-        &self,
-        aggregator: &Aggregator,
-        hash: &str,
-        expected_epoch_min: Epoch,
-    ) -> StdResult<String> {
-        let url = format!(
-            "{}/artifact/cardano-transaction/{hash}",
-            aggregator.endpoint()
-        );
-        info!(
-            "Asserting the aggregator is signing the Cardano transactions artifact `{}` with an expected min epoch of `{}`",
-            hash,
-            expected_epoch_min;
-            "aggregator" => &aggregator.name()
-        );
-
-        async fn fetch_cardano_transaction_snapshot_message(
-            url: String,
-            expected_epoch_min: Epoch,
-        ) -> StdResult<Option<CardanoTransactionSnapshotMessage>> {
-            match get_json_response::<CardanoTransactionSnapshotMessage>(url).await? {
-                Ok(artifact) => match artifact.epoch {
-                    epoch if epoch >= expected_epoch_min => Ok(Some(artifact)),
-                    epoch => Err(anyhow!(
-                        "Minimum expected artifact epoch not reached: {epoch} < {expected_epoch_min}"
-                    )),
-                },
-                Err(err) => Err(anyhow!(err).context("Invalid Cardano transactions artifact body")),
-            }
-        }
-
-        match attempt!(10, Duration::from_millis(1000), {
-        fetch_cardano_transaction_snapshot_message(url.clone(), expected_epoch_min).await
-    }) {
-            AttemptResult::Ok(artifact) => {
-                info!("Signer signed a Cardano transactions artifact"; "certificate_hash" => &artifact.certificate_hash, "aggregator" => &aggregator.name());
-                Ok(artifact.certificate_hash)
-            }
-            AttemptResult::Err(error) => Err(error),
-            AttemptResult::Timeout() => Err(anyhow!(
-            "Timeout exhausted assert_signer_is_signing_cardano_transactions, no response from `{url}`"
-        )),
-        }.with_context(|| {
-            format!(
-                "Requesting aggregator `{}`",
-                aggregator.name()
-            )
-        })
-    }
-
-    pub async fn node_producing_cardano_blocks_transactions(
-        &self,
-        aggregator: &Aggregator,
-    ) -> StdResult<String> {
-        let url = format!(
-            "{}/artifact/cardano-blocks-transactions",
-            aggregator.endpoint()
-        );
-        info!("Waiting for the aggregator to produce a Cardano blocks transactions artifact"; "aggregator" => &aggregator.name());
-
-        async fn fetch_last_cardano_blocks_transactions_snapshot_hash(
-            url: String,
-        ) -> StdResult<Option<String>> {
-            match get_json_response::<CardanoBlocksTransactionsSnapshotListMessage>(url)
-                .await?
-                .as_deref()
-            {
-                Ok([artifact, ..]) => Ok(Some(artifact.hash.clone())),
-                Ok(&[]) => Ok(None),
-                Err(err) => Err(anyhow!(
-                    "Invalid Cardano blocks transactions artifact body: {err}",
-                )),
-            }
-        }
-
-        match attempt!(30, Duration::from_millis(2000), {
-        fetch_last_cardano_blocks_transactions_snapshot_hash(url.clone()).await
-    }) {
-            AttemptResult::Ok(hash) => {
-                info!("Aggregator produced a Cardano blocks transactions artifact"; "hash" => &hash, "aggregator" => &aggregator.name());
-                Ok(hash)
-            }
-            AttemptResult::Err(error) => Err(error),
-            AttemptResult::Timeout() => Err(anyhow!(
-            "Timeout exhausted assert_node_producing_cardano_blocks_transactions, no response from `{url}`"
-        )),
-        }
-            .with_context(|| format!("Requesting aggregator `{}`", aggregator.name()))
-    }
-
-    pub async fn signer_is_signing_cardano_blocks_transactions(
-        &self,
-        aggregator: &Aggregator,
-        hash: &str,
-        expected_epoch_min: Epoch,
-    ) -> StdResult<String> {
-        let url = format!(
-            "{}/artifact/cardano-blocks-transactions/{hash}",
-            aggregator.endpoint()
-        );
-        info!(
-            "Asserting the aggregator is signing the Cardano blocks transactions artifact `{}` with an expected min epoch of `{}`",
-            hash,
-            expected_epoch_min;
-            "aggregator" => &aggregator.name()
-        );
-
-        async fn fetch_cardano_blocks_transactions_snapshot_message(
-            url: String,
-            expected_epoch_min: Epoch,
-        ) -> StdResult<Option<CardanoBlocksTransactionsSnapshotMessage>> {
-            match get_json_response::<CardanoBlocksTransactionsSnapshotMessage>(url).await? {
-                Ok(artifact) => match artifact.epoch {
-                    epoch if epoch >= expected_epoch_min => Ok(Some(artifact)),
-                    epoch => Err(anyhow!(
-                        "Minimum expected artifact epoch not reached: {epoch} < {expected_epoch_min}"
-                    )),
-                },
-                Err(err) => {
-                    Err(anyhow!(err).context("Invalid Cardano blocks transactions artifact body"))
-                }
-            }
-        }
-
-        match attempt!(10, Duration::from_millis(1000), {
-        fetch_cardano_blocks_transactions_snapshot_message(url.clone(), expected_epoch_min).await
-    }) {
-            AttemptResult::Ok(artifact) => {
-                info!("Signer signed a Cardano blocks transactions artifact"; "certificate_hash" => &artifact.certificate_hash, "aggregator" => &aggregator.name());
-                Ok(artifact.certificate_hash)
-            }
-            AttemptResult::Err(error) => Err(error),
-            AttemptResult::Timeout() => Err(anyhow!(
-            "Timeout exhausted assert_signer_is_signing_cardano_blocks_transactions, no response from `{url}`"
-        )),
-        }.with_context(|| {
-            format!(
-                "Requesting aggregator `{}`",
-                aggregator.name()
-            )
-        })
-    }
-
-    pub async fn node_producing_cardano_stake_distribution(
-        &self,
-        aggregator: &Aggregator,
-    ) -> StdResult<(String, Epoch)> {
-        let url = format!(
-            "{}/artifact/cardano-stake-distributions",
-            aggregator.endpoint()
-        );
-        info!("Waiting for the aggregator to produce a Cardano stake distribution"; "aggregator" => &aggregator.name());
-
-        async fn fetch_last_cardano_stake_distribution_message(
-            url: String,
-        ) -> StdResult<Option<(String, Epoch)>> {
-            match get_json_response::<CardanoStakeDistributionListMessage>(url)
-                .await?
-                .as_deref()
-            {
-                Ok([stake_distribution, ..]) => Ok(Some((
-                    stake_distribution.hash.clone(),
-                    stake_distribution.epoch,
-                ))),
-                Ok(&[]) => Ok(None),
-                Err(err) => Err(anyhow!("Invalid Cardano stake distribution body: {err}",)),
-            }
-        }
-
-        match attempt!(30, Duration::from_millis(2000), {
-        fetch_last_cardano_stake_distribution_message(url.clone()).await
-    }) {
-            AttemptResult::Ok((hash, epoch)) => {
-                info!("Aggregator produced a Cardano stake distribution"; "hash" => &hash, "epoch" => #?epoch, "aggregator" => &aggregator.name());
-                Ok((hash, epoch))
-            }
-            AttemptResult::Err(error) => Err(error),
-            AttemptResult::Timeout() => Err(anyhow!(
-            "Timeout exhausted assert_node_producing_cardano_stake_distribution, no response from `{url}`"
-        )),
-        }.with_context(|| {
-            format!(
-                "Requesting aggregator `{}`",
-                aggregator.name()
-            )
-        })
-    }
-
-    pub async fn signer_is_signing_cardano_stake_distribution(
-        &self,
-        aggregator: &Aggregator,
-        hash: &str,
-        expected_epoch_min: Epoch,
-    ) -> StdResult<String> {
-        let url = format!(
-            "{}/artifact/cardano-stake-distribution/{hash}",
-            aggregator.endpoint()
-        );
-        info!(
-            "Asserting the aggregator is signing the Cardano stake distribution message `{}` with an expected min epoch of `{}`",
-            hash,
-            expected_epoch_min;
-            "aggregator" => &aggregator.name()
-        );
-
-        async fn fetch_cardano_stake_distribution_message(
-            url: String,
-            expected_epoch_min: Epoch,
-        ) -> StdResult<Option<CardanoStakeDistributionMessage>> {
-            match get_json_response::<CardanoStakeDistributionMessage>(url).await? {
-                Ok(stake_distribution) => match stake_distribution.epoch {
-                    epoch if epoch >= expected_epoch_min => Ok(Some(stake_distribution)),
-                    epoch => Err(anyhow!(
-                        "Minimum expected Cardano stake distribution epoch not reached: {epoch} < {expected_epoch_min}"
-                    )),
-                },
-                Err(err) => Err(anyhow!(err).context("Invalid Cardano stake distribution body")),
-            }
-        }
-
-        match attempt!(10, Duration::from_millis(1000), {
-        fetch_cardano_stake_distribution_message(url.clone(), expected_epoch_min).await
-    }) {
-            AttemptResult::Ok(cardano_stake_distribution) => {
-                info!("Signer signed a Cardano stake distribution"; "certificate_hash" => &cardano_stake_distribution.certificate_hash, "aggregator" => &aggregator.name());
-                Ok(cardano_stake_distribution.certificate_hash)
-            }
-            AttemptResult::Err(error) => Err(error),
-            AttemptResult::Timeout() => Err(anyhow!(
-            "Timeout exhausted assert_signer_is_signing_cardano_stake_distribution, no response from `{url}`"
-        )),
-        }.with_context(|| {
-            format!(
-                "Requesting aggregator `{}`",
-                aggregator.name()
-            )
-        })
-    }
-
-    pub async fn is_creating_certificate_with_enough_signers(
-        &self,
-        aggregator: &Aggregator,
-        certificate_hash: &str,
-        total_signers_expected: usize,
-    ) -> StdResult<()> {
-        let url = format!("{}/certificate/{certificate_hash}", aggregator.endpoint());
-        info!("Waiting for the aggregator to create a certificate with enough signers"; "aggregator" => &aggregator.name());
-
-        async fn fetch_certificate_message(url: String) -> StdResult<Option<CertificateMessage>> {
-            match get_json_response::<CertificateMessage>(url).await? {
-                Ok(certificate) => Ok(Some(certificate)),
-                Err(err) => Err(anyhow!(err).context("Invalid snapshot body")),
-            }
-        }
-
-        match attempt!(10, Duration::from_millis(1000), {
-            fetch_certificate_message(url.clone()).await
-        }) {
-            AttemptResult::Ok(certificate) => {
-                info!("Aggregator produced a certificate"; "certificate" => ?certificate);
-                if certificate.metadata.signers.len() == total_signers_expected {
-                    info!(
-                        "Certificate is signed by expected number of signers: {} >= {} ",
-                        certificate.metadata.signers.len(),
-                        total_signers_expected ;
-                        "aggregator" => &aggregator.name()
-                    );
-                    Ok(())
-                } else {
-                    Err(anyhow!(
-                        "Certificate is not signed by expected number of signers: {} < {} ",
-                        certificate.metadata.signers.len(),
-                        total_signers_expected
-                    ))
-                }
-            }
-            AttemptResult::Err(error) => Err(error),
-            AttemptResult::Timeout() => Err(anyhow!(
-                "Timeout exhausted assert_is_creating_certificate, no response from `{url}`"
-            )),
-        }
-        .with_context(|| format!("Requesting aggregator `{}`", aggregator.name()))
-    }
-
-    pub async fn client_can_verify_cardano_database(
-        &self,
-        client: &mut Client,
-        hash: &str,
-    ) -> StdResult<()> {
-        client
-            .run(ClientCommand::CardanoDbV2(CardanoDbV2Command::List))
-            .await?;
-
-        if client.version().is_above_or_equal("0.12.34") {
-            client
-                .run(ClientCommand::CardanoDbV2(
-                    CardanoDbV2Command::ListPerEpoch {
-                        epoch_specifier: EpochSpecifier::LatestMinusOffset(5),
-                    },
-                ))
-                .await?;
-        } else {
-            warn!(
-                "Client version is below 0.12.34, skipping `cardano-db snapshot list --epoch latest-5` check"
-            );
-        }
-
-        client
-            .run(ClientCommand::CardanoDbV2(CardanoDbV2Command::Show {
-                hash: hash.to_string(),
-            }))
-            .await?;
-        info!("Client list & show the cardano database snapshot"; "hash" => &hash);
-
-        client
-            .run(ClientCommand::CardanoDbV2(CardanoDbV2Command::Download {
-                hash: hash.to_string(),
-            }))
-            .await?;
-        info!("Client downloaded & restored the cardano database snapshot"; "hash" => &hash);
-
-        Ok(())
-    }
-
-    pub async fn client_can_verify_mithril_stake_distribution(
-        &self,
-        client: &mut Client,
-        hash: &str,
-    ) -> StdResult<()> {
-        client
-            .run(ClientCommand::MithrilStakeDistribution(
-                MithrilStakeDistributionCommand::Download {
-                    hash: hash.to_owned(),
-                },
-            ))
-            .await?;
-        info!("Client downloaded the Mithril stake distribution"; "hash" => &hash);
-
-        Ok(())
-    }
-
-    pub async fn client_can_verify_transactions(
-        &self,
-        client: &mut Client,
-        tx_hashes: Vec<TransactionHash>,
-    ) -> StdResult<()> {
-        #[allow(dead_code)]
-        #[derive(Debug, serde::Deserialize)]
-        struct ClientCardanoTransactionCertifyResult {
-            certified_transactions: Vec<TransactionHash>,
-            non_certified_transactions: Vec<TransactionHash>,
-        }
-
-        let result_file = client
-            .run(ClientCommand::CardanoTransaction(
-                CardanoTransactionCommand::Certify {
-                    tx_hashes: tx_hashes.clone(),
-                },
-            ))
-            .await?;
-        info!("Client verified the Cardano transactions"; "tx_hashes" => ?tx_hashes);
-
-        let file = std::fs::read_to_string(&result_file).with_context(|| {
-            format!(
-                "Failed to read client output from file `{}`",
-                result_file.display()
-            )
-        })?;
-        let result: ClientCardanoTransactionCertifyResult = serde_json::from_str(&file)
-            .with_context(|| {
-                format!(
-                    "Failed to parse client output as json from file `{}`",
-                    result_file.display()
-                )
-            })?;
-
-        info!("Asserting that all Cardano transactions were verified by the Client...");
-        if tx_hashes.iter().all(|tx| result.certified_transactions.contains(tx)) {
-            Ok(())
-        } else {
-            Err(anyhow!(
-                "Not all transactions were certified:\n'{:#?}'",
-                result,
-            ))
-        }
-    }
-
-    pub async fn client_can_verify_transactions_v2(
-        &self,
-        client: &mut Client,
-        tx_hashes: Vec<TransactionHash>,
-    ) -> StdResult<()> {
-        #[allow(dead_code)]
-        #[derive(Debug, serde::Deserialize)]
-        struct ClientCardanoTransactionCertifyResult {
-            certified_transactions: Vec<CertifiedTransactionV2>,
-            non_certified_transactions: Vec<TransactionHash>,
-        }
-
-        #[derive(Debug, serde::Deserialize)]
-        struct CertifiedTransactionV2 {
-            transaction_hash: String,
-        }
-
-        if !client.version().is_above_or_equal("0.13.1") {
-            warn!(
-                "Client version is below 0.13.1, skipping `cardano-transaction certify --backend v2` check"
-            );
-            return Ok(());
-        }
-
-        let result_file = client
-            .run(ClientCommand::CardanoTransactionV2(
-                CardanoTransactionV2Command::Certify {
-                    tx_hashes: tx_hashes.clone(),
-                },
-            ))
-            .await?;
-        info!("Client verified the Cardano transactions V2"; "tx_hashes" => ?tx_hashes);
-
-        let file = std::fs::read_to_string(&result_file).with_context(|| {
-            format!(
-                "Failed to read client output from file `{}`",
-                result_file.display()
-            )
-        })?;
-        let result: ClientCardanoTransactionCertifyResult = serde_json::from_str(&file)
-            .with_context(|| {
-                format!(
-                    "Failed to parse client output as json from file `{}`",
-                    result_file.display()
-                )
-            })?;
-
-        info!("Asserting that all Cardano transactions V2 were verified by the Client...");
-        let certified_tx_hashes_result: Vec<String> = result
-            .certified_transactions
-            .iter()
-            .map(|tx| tx.transaction_hash.clone())
-            .collect();
-
-        if tx_hashes.iter().all(|tx| certified_tx_hashes_result.contains(tx)) {
-            Ok(())
-        } else {
-            Err(anyhow!(
-                "Not all transactions V2 were certified:\n'{:#?}'",
-                result,
-            ))
-        }
-    }
-
-    pub async fn client_can_verify_blocks(
-        &self,
-        client: &mut Client,
-        block_hashes: Vec<BlockHash>,
-    ) -> StdResult<()> {
-        #[allow(dead_code)]
-        #[derive(Debug, serde::Deserialize)]
-        struct ClientCardanoBlockCertifyResult {
-            certified_blocks: Vec<CertifiedBlock>,
-            non_certified_blocks: Vec<BlockHash>,
-        }
-
-        #[derive(Debug, serde::Deserialize)]
-        struct CertifiedBlock {
-            block_hash: String,
-        }
-
-        if !client.version().is_above_or_equal("0.13.1") {
-            warn!("Client version is below 0.13.1, skipping `cardano-block certify` check");
-            return Ok(());
-        }
-
-        let result_file = client
-            .run(ClientCommand::CardanoBlock(CardanoBlockCommand::Certify {
-                block_hashes: block_hashes.clone(),
-            }))
-            .await?;
-        info!("Client verified the Cardano blocks"; "block_hashes" => ?block_hashes);
-        let file = std::fs::read_to_string(&result_file).with_context(|| {
-            format!(
-                "Failed to read client output from file `{}`",
-                result_file.display()
-            )
-        })?;
-        let result: ClientCardanoBlockCertifyResult =
-            serde_json::from_str(&file).with_context(|| {
-                format!(
-                    "Failed to parse client output as json from file `{}`",
-                    result_file.display()
-                )
-            })?;
-
-        info!("Asserting that all Cardano blocks were verified by the Client...");
-        let certified_blocks_hashes_result: Vec<String> = result
-            .certified_blocks
-            .iter()
-            .map(|block| block.block_hash.clone())
-            .collect();
-
-        if block_hashes
-            .iter()
-            .all(|block| certified_blocks_hashes_result.contains(block))
-        {
-            Ok(())
-        } else {
-            Err(anyhow!("Not all blocks were certified:\n'{:#?}'", result,))
-        }
-    }
-
-    pub async fn client_can_verify_cardano_stake_distribution(
-        &self,
-        client: &mut Client,
-        hash: &str,
-        epoch: Epoch,
-    ) -> StdResult<()> {
-        client
-            .run(ClientCommand::CardanoStakeDistribution(
-                CardanoStakeDistributionCommand::Download {
-                    unique_identifier: epoch.to_string(),
-                },
-            ))
-            .await?;
-        info!("Client downloaded the Cardano stake distribution by epoch"; "epoch" => epoch.to_string());
-
-        client
-            .run(ClientCommand::CardanoStakeDistribution(
-                CardanoStakeDistributionCommand::Download {
-                    unique_identifier: hash.to_string(),
-                },
-            ))
-            .await?;
-        info!("Client downloaded the Cardano stake distribution by hash"; "hash" => hash.to_string());
-
-        Ok(())
     }
 
     pub async fn client_can_convert_the_ledger_snapshot(
@@ -898,6 +134,7 @@ pub async fn assert_node_producing_mithril_stake_distribution(
     aggregator: &Aggregator,
 ) -> StdResult<String> {
     CheckToolkit::default()
+        .mithril_stake_distribution
         .node_producing_mithril_stake_distribution(aggregator)
         .await
 }
@@ -908,6 +145,7 @@ pub async fn assert_signer_is_signing_mithril_stake_distribution(
     expected_epoch_min: Epoch,
 ) -> StdResult<String> {
     CheckToolkit::default()
+        .mithril_stake_distribution
         .signer_is_signing_mithril_stake_distribution(aggregator, hash, expected_epoch_min)
         .await
 }
@@ -916,6 +154,7 @@ pub async fn assert_node_producing_cardano_database_snapshot(
     aggregator: &Aggregator,
 ) -> StdResult<String> {
     CheckToolkit::default()
+        .cardano_database
         .node_producing_cardano_database_snapshot(aggregator)
         .await
 }
@@ -926,6 +165,7 @@ pub async fn assert_signer_is_signing_cardano_database_snapshot(
     expected_epoch_min: Epoch,
 ) -> StdResult<String> {
     CheckToolkit::default()
+        .cardano_database
         .signer_is_signing_cardano_database_snapshot(aggregator, hash, expected_epoch_min)
         .await
 }
@@ -934,6 +174,7 @@ pub async fn assert_node_producing_cardano_database_digests_map(
     aggregator: &Aggregator,
 ) -> StdResult<Vec<(String, String)>> {
     CheckToolkit::default()
+        .cardano_database
         .node_producing_cardano_database_digests_map(aggregator)
         .await
 }
@@ -942,6 +183,7 @@ pub async fn assert_node_producing_cardano_transactions(
     aggregator: &Aggregator,
 ) -> StdResult<String> {
     CheckToolkit::default()
+        .cardano_transactions
         .node_producing_cardano_transactions(aggregator)
         .await
 }
@@ -952,6 +194,7 @@ pub async fn assert_signer_is_signing_cardano_transactions(
     expected_epoch_min: Epoch,
 ) -> StdResult<String> {
     CheckToolkit::default()
+        .cardano_transactions
         .signer_is_signing_cardano_transactions(aggregator, hash, expected_epoch_min)
         .await
 }
@@ -960,6 +203,7 @@ pub async fn assert_node_producing_cardano_blocks_transactions(
     aggregator: &Aggregator,
 ) -> StdResult<String> {
     CheckToolkit::default()
+        .cardano_blocks_transactions
         .node_producing_cardano_blocks_transactions(aggregator)
         .await
 }
@@ -970,6 +214,7 @@ pub async fn assert_signer_is_signing_cardano_blocks_transactions(
     expected_epoch_min: Epoch,
 ) -> StdResult<String> {
     CheckToolkit::default()
+        .cardano_blocks_transactions
         .signer_is_signing_cardano_blocks_transactions(aggregator, hash, expected_epoch_min)
         .await
 }
@@ -978,6 +223,7 @@ pub async fn assert_node_producing_cardano_stake_distribution(
     aggregator: &Aggregator,
 ) -> StdResult<(String, Epoch)> {
     CheckToolkit::default()
+        .cardano_stake_distribution
         .node_producing_cardano_stake_distribution(aggregator)
         .await
 }
@@ -988,6 +234,7 @@ pub async fn assert_signer_is_signing_cardano_stake_distribution(
     expected_epoch_min: Epoch,
 ) -> StdResult<String> {
     CheckToolkit::default()
+        .cardano_stake_distribution
         .signer_is_signing_cardano_stake_distribution(aggregator, hash, expected_epoch_min)
         .await
 }
@@ -998,6 +245,7 @@ pub async fn assert_is_creating_certificate_with_enough_signers(
     total_signers_expected: usize,
 ) -> StdResult<()> {
     CheckToolkit::default()
+        .certificate
         .is_creating_certificate_with_enough_signers(
             aggregator,
             certificate_hash,
@@ -1011,6 +259,7 @@ pub async fn assert_client_can_verify_cardano_database(
     hash: &str,
 ) -> StdResult<()> {
     CheckToolkit::default()
+        .cardano_database
         .client_can_verify_cardano_database(client, hash)
         .await
 }
@@ -1020,6 +269,7 @@ pub async fn assert_client_can_verify_mithril_stake_distribution(
     hash: &str,
 ) -> StdResult<()> {
     CheckToolkit::default()
+        .mithril_stake_distribution
         .client_can_verify_mithril_stake_distribution(client, hash)
         .await
 }
@@ -1029,6 +279,7 @@ pub async fn assert_client_can_verify_transactions(
     tx_hashes: Vec<TransactionHash>,
 ) -> StdResult<()> {
     CheckToolkit::default()
+        .cardano_transactions
         .client_can_verify_transactions(client, tx_hashes)
         .await
 }
@@ -1038,7 +289,8 @@ pub async fn assert_client_can_verify_transactions_v2(
     tx_hashes: Vec<TransactionHash>,
 ) -> StdResult<()> {
     CheckToolkit::default()
-        .client_can_verify_transactions_v2(client, tx_hashes)
+        .cardano_blocks_transactions
+        .client_verify_transactions(client, tx_hashes)
         .await
 }
 
@@ -1047,7 +299,8 @@ pub async fn assert_client_can_verify_blocks(
     block_hashes: Vec<BlockHash>,
 ) -> StdResult<()> {
     CheckToolkit::default()
-        .client_can_verify_blocks(client, block_hashes)
+        .cardano_blocks_transactions
+        .client_verify_blocks(client, block_hashes)
         .await
 }
 
@@ -1057,6 +310,7 @@ pub async fn assert_client_can_verify_cardano_stake_distribution(
     epoch: Epoch,
 ) -> StdResult<()> {
     CheckToolkit::default()
+        .cardano_stake_distribution
         .client_can_verify_cardano_stake_distribution(client, hash, epoch)
         .await
 }
